convert-video/requirements.txt
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.2.3
SQLAlchemy==1.4.15
Werkzeug==2.0.1
psycopg2-binary
flask_migrate
flask
flask_cors
aiven-client
python-dotenv
eventlet
Flask-Executor
google-cloud-storage
